Move physics table conversion to a free function
