+++ src/celeritas/CMakeLists.txt
@@ -101,6 +101,7 @@ if(CELERITAS_USE_Geant4)
     ext/detail/GeantBremsstrahlungProcess.cc
     ext/detail/GeantExceptionHandler.cc
     ext/detail/GeantGeoExporter.cc
+    ext/detail/GeantImportUtils.cc
     ext/detail/GeantLoggerAdapter.cc
     ext/detail/GeantPhysicsList.cc
     ext/detail/ImportProcessConverter.cc
