+++ src/celeritas/ext/detail/ImportProcessConverter.cc
@@ -23,9 +23,7 @@
 #include <G4NistManager.hh>
 #include <G4ParticleDefinition.hh>
 #include <G4ParticleTable.hh>
-#include <G4PhysicsTable.hh>
 #include <G4PhysicsVector.hh>
-#include <G4PhysicsVectorType.hh>
 #include <G4ProcessType.hh>
 #include <G4ProductionCutsTable.hh>
 #include <G4String.hh>
@@ -44,6 +42,7 @@
 #include "celeritas/phys/PDGNumber.hh"
 
 #include "../GeantConfig.hh"
+#include "GeantImportUtils.hh"
 
 using CLHEP::cm;
 using CLHEP::cm2;
@@ -154,63 +153,6 @@ ImportModelClass to_import_model(G4VEmModel const& model)
         return ImportModelClass::none;
     }
     return iter->second;
-}
-
-//---------------------------------------------------------------------------//
-/*!
- * Safely switch from \c G4PhysicsVectorType to \c ImportPhysicsVectorType .
- * [See G4PhysicsVectorType.hh]
- *
- * Geant4 v11 has a different set of G4PhysicsVectorType enums.
- */
-ImportPhysicsVectorType
-to_import_physics_vector_type(G4PhysicsVectorType g4_vector_type)
-{
-    switch (g4_vector_type)
-    {
-#if CELERITAS_G4_V10
-        case T_G4PhysicsVector:
-            return ImportPhysicsVectorType::unknown;
-#endif
-        case T_G4PhysicsLinearVector:
-            return ImportPhysicsVectorType::linear;
-        case T_G4PhysicsLogVector:
-#if CELERITAS_G4_V10
-        case T_G4PhysicsLnVector:
-#endif
-            return ImportPhysicsVectorType::log;
-        case T_G4PhysicsFreeVector:
-#if CELERITAS_G4_V10
-        case T_G4PhysicsOrderedFreeVector:
-        case T_G4LPhysicsFreeVector:
-#endif
-            return ImportPhysicsVectorType::free;
-    }
-    CELER_ASSERT_UNREACHABLE();
-}
-
-//---------------------------------------------------------------------------//
-/*!
- * Get a multiplicative geant-natural-units constant to convert the units.
- */
-double units_to_scaling(ImportUnits units)
-{
-    switch (units)
-    {
-        case ImportUnits::none:
-            return 1;
-        case ImportUnits::cm_inv:
-            return cm;
-        case ImportUnits::cm_mev_inv:
-            return cm * MeV;
-        case ImportUnits::mev:
-            return 1 / MeV;
-        case ImportUnits::mev_per_cm:
-            return cm / MeV;
-        case ImportUnits::cm:
-            return 1 / cm;
-    }
-    CELER_ASSERT_UNREACHABLE();
 }
 
 //---------------------------------------------------------------------------//
@@ -547,64 +489,7 @@ void ImportProcessConverter::add_table(G4PhysicsTable const* g4table,
         return;
     }
 
-    ImportPhysicsTable table;
-    table.table_type = table_type;
-    switch (table_type)
-    {
-        case ImportTableType::dedx:
-        case ImportTableType::dedx_process:
-        case ImportTableType::dedx_subsec:
-        case ImportTableType::dedx_unrestricted:
-        case ImportTableType::ionization:
-        case ImportTableType::ionization_subsec:
-            table.x_units = ImportUnits::mev;
-            table.y_units = ImportUnits::mev_per_cm;
-            break;
-        case ImportTableType::csda_range:
-        case ImportTableType::range:
-        case ImportTableType::secondary_range:
-            table.x_units = ImportUnits::mev;
-            table.y_units = ImportUnits::cm;
-            break;
-        case ImportTableType::inverse_range:
-            table.x_units = ImportUnits::cm;
-            table.y_units = ImportUnits::mev;
-            break;
-        case ImportTableType::lambda:
-        case ImportTableType::sublambda:
-            table.x_units = ImportUnits::mev;
-            table.y_units = ImportUnits::cm_inv;
-            break;
-        case ImportTableType::lambda_prim:
-            table.x_units = ImportUnits::mev;
-            table.y_units = ImportUnits::cm_mev_inv;
-            break;
-        default:
-            CELER_ASSERT_UNREACHABLE();
-    };
-
-    // Convert units
-    double x_scaling = units_to_scaling(table.x_units);
-    double y_scaling = units_to_scaling(table.y_units);
-
-    // Save physics vectors
-    for (auto const* g4vector : *g4table)
-    {
-        ImportPhysicsVector import_vec;
-
-        // Populate ImportPhysicsVectors
-        import_vec.vector_type
-            = to_import_physics_vector_type(g4vector->GetType());
-
-        for (auto j : celeritas::range(g4vector->GetVectorLength()))
-        {
-            import_vec.x.push_back(g4vector->Energy(j) * x_scaling);
-            import_vec.y.push_back((*g4vector)[j] * y_scaling);
-        }
-        table.physics_vectors.push_back(std::move(import_vec));
-    }
-
-    process_.tables.push_back(std::move(table));
+    process_.tables.push_back(import_table(*g4table, table_type));
 }
 
 //---------------------------------------------------------------------------//
